requirements.txt
a2a-sdk==0.3.7
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.16.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asteval==1.0.6
attrs==25.4.0
Authlib==1.6.5
azure-ai-documentintelligence==1.0.2
azure-cognitiveservices-speech==1.41.1
azure-core==1.38.0
azure-identity==1.25.1
babel==2.18.0
beautifulsoup4==4.13.5
blinker==1.9.0
boto3==1.40.37
botocore==1.40.76
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cloudpickle==3.1.2
cobble==0.1.4
colorama==0.4.6
coloredlogs==15.0.1
cryptography==46.0.4
defusedxml==0.7.1
distro==1.9.0
docstring_parser==0.17.0
et_xmlfile==2.0.0
fastapi==0.120.1
fastuuid==0.14.0
filelock==3.20.3
Flask==3.0.3
flask-cors==6.0.1
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.1.0
gevent==25.8.2
gevent-websocket==0.10.1
gitdb==4.0.12
GitPython==3.1.45
google-adk==1.18.0
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-cloud-aiplatform==1.133.0
google-cloud-appengine-logging==1.8.0
google-cloud-audit-log==0.4.0
google-cloud-bigquery==3.40.0
google-cloud-bigtable==2.35.0
google-cloud-core==2.5.0
google-cloud-discoveryengine==0.13.12
google-cloud-firestore==2.23.0
google-cloud-logging==3.13.0
google-cloud-monitoring==2.29.0
google-cloud-resource-manager==1.16.0
google-cloud-secret-manager==2.26.0
google-cloud-spanner==3.62.0
google-cloud-speech==2.36.0
google-cloud-storage==3.9.0
google-cloud-trace==1.18.0
google-crc32c==1.8.0
google-genai==1.49.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
graphviz==0.21
greenlet==3.3.1
grpc-google-iam-v1==0.14.3
grpc-interceptor==0.15.4
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
hf-xet==1.2.0
holidays==0.81
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.3.7
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.1
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.0.1
jsonpath-ng==1.7.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jwcrypto==1.5.6
kaleido==0.2.1
litellm==1.76.3
lxml==6.0.2
magika==0.6.3
Mako==1.3.10
mammoth==1.11.0
markdown-it-py==4.0.0
markdownify==1.2.0
markitdown==0.1.4
MarkupSafe==3.0.3
mcp==1.26.0
mdurl==0.1.2
mermaid_cli==0.1.2
mmh3==5.2.0
mpmath==1.3.0
msal==1.34.0
msal-extensions==1.3.1
multidict==6.7.1
narwhals==2.16.0
numpy==2.2.6
olefile==0.47
onnxruntime==1.20.1
openai==1.99.9
openpyxl==3.1.5
opentelemetry-api==1.37.0
opentelemetry-exporter-gcp-logging==1.11.0a0
opentelemetry-exporter-gcp-monitoring==1.11.0a0
opentelemetry-exporter-gcp-trace==1.11.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-resourcedetector-gcp==1.11.0a0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
packaging==26.0
pandas==2.3.2
pdfminer.six==20260107
pillow==12.1.0
playwright==1.55.0
plotly==6.3.0
ply==3.11
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
psycopg2-binary==2.9.10
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.11.9
pydantic-settings==2.12.0
pydantic_core==2.33.2
pydub==0.25.1
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.11.0
pyodbc==5.3.0
pyparsing==3.3.2
pyreadline3==3.5.4
pystache==0.6.8
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==4.0.0
python-jwt==4.1.0
python-liquid==2.1.0
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
pywin32==311
PyYAML==6.0.2
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rich==13.9.4
rouge==1.0.1
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.14.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
solace-agent-mesh==1.14.0
solace-pubsubplus==1.9.0
solace_ai_connector==3.3.0
soupsieve==2.8.3
SpeechRecognition==3.14.5
SQLAlchemy==2.0.40
sqlalchemy-spanner==1.17.2
sqlparse==0.5.5
sse-starlette==3.0.2
starlette==0.49.1
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.2
toml==0.10.2
tqdm==4.67.2
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.37.0
watchdog==6.0.0
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.5
xlrd==2.0.2
xlsxwriter==3.2.9
yarl==1.22.0
youtube-transcript-api==1.0.3
zipp==3.23.0
zope.event==6.1
zope.interface==8.2
